Add figures 8 and 9; update to-do list to reflect
Finished the last figures that needed to be converted to SVG! There are
some figures in the paper that I won't recreate, as they are screenshots
from TensorBoard and EEG. The new figures are added to README.md, and
the to-do list has removed those features.

Right now, the last big thing to do is add a whole bunch of reference
links around the page to aid in navigation. Probably add a table of
contents to fit it.